add not supported warning in autosave

+++ chemcanvas/managers.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2026 Arindam Chaudhuri <arindamsoft94@gmail.com>
 import os
 import csv, glob
-from PyQt5.QtCore import QObject, QTimer, Qt, QSettings, QLockFile, QStandardPaths
+from PyQt5.QtCore import QObject, QTimer, Qt, QSettings, QLockFile
 from PyQt5.QtWidgets import (QMessageBox, QDialog, QGridLayout, QCheckBox,
         QLabel, QSpinBox, QDialogButtonBox)
 from app_data import App, Settings
@@ -21,17 +21,20 @@ class AutosaveManager(QObject):
     def __init__(self, window):
         super().__init__(window)
         self.window = window
+        if not self.is_supported:
+            return
         self.autosave_dir = App.DATA_DIR + "/autosaves"
         os.makedirs(self.autosave_dir, exist_ok=True)
         self.id = str(os.getpid())
-        self.lockfile = None
-        self.lock_dir = self.autosave_dir
-        if os.path.exists("/.flatpak-info"):
-            self.lock_dir = QStandardPaths.writableLocation(QStandardPaths.RuntimeLocation)
+        self._lock = None
         # timer that triggers autosave_all
         self.timer = QTimer(self)
         self.timer.timeout.connect(self.autosave_all)
         self.timer.start(Settings.autosave_interval * 1000)
+
+    @property
+    def is_supported(self):
+        return not os.path.exists("/.flatpak-info")
 
 
     def autosave_all(self):
@@ -62,12 +65,11 @@ class AutosaveManager(QObject):
                 writer.writerows(metadata)
                 # create a lockfile to ensure other instance of this program
                 # knows if it is being used, and does not prompt for restore
-                lock_path = self.lock_dir + f"/autosaves-{self.id}.lock"
-                self.lockfile = QLockFile(lock_path)
-                self.lockfile.setStaleLockTime(0)
-                if not self.lockfile.tryLock(0):
-                    err_code = self.lockfile.error()
-                    print(f"Error {err_code} : could not lock file {lock_path}")
+                lock_path = self.autosave_dir + f"/autosaves-{self.id}.lock"
+                self._lock = QLockFile(lock_path)
+                self._lock.setStaleLockTime(0)
+                if not self._lock.tryLock(0):
+                    print("could not lock file", lock_path)
         except Exception as e:
             print(e)
 
@@ -84,6 +86,8 @@ class AutosaveManager(QObject):
 
 
     def remove_backup_for_tab(self, tab):
+        if not self.is_supported:
+            return
         filepath = self.autosave_dir + f"/{self.id}-{tab.id}.ccdx"
         try:
             os.remove(filepath)
@@ -92,6 +96,8 @@ class AutosaveManager(QObject):
 
 
     def remove_all_backups(self):
+        if not self.is_supported:
+            return
         self.remove_backups_for_id(self.id)
 
     def remove_backups_for_id(self, uid):
@@ -99,9 +105,9 @@ class AutosaveManager(QObject):
             os.remove(self.autosave_dir + f"/autosaves-{uid}.csv")
             # release the lockfile
             if uid==self.id:
-                self.lockfile.unlock()
+                self._lock.unlock()
             else:
-                lockfile = QLockFile(self.lock_dir + f"/autosaves-{uid}.lock")
+                lockfile = QLockFile(self.autosave_dir + f"/autosaves-{uid}.lock")
                 lockfile.unlock()
         except:
             pass
@@ -115,6 +121,8 @@ class AutosaveManager(QObject):
     def check_and_offer_restore(self):
         """ Called at startup: if autosaves exist, offer to restore them.
         returns True if restored """
+        if not self.is_supported:
+            return
         metadata = [] # list of [uid, tab_id, unsaved, filepath]
         # find autosaves
         files = [f for f in glob.glob(self.autosave_dir + "/autosaves-*.csv")]
@@ -123,7 +131,7 @@ class AutosaveManager(QObject):
             uid = basename[10:-4]
             if uid==self.id:
                 continue
-            lockfile = QLockFile(self.lock_dir + f"/autosaves-{uid}.lock")
+            lockfile = QLockFile(self.autosave_dir + f"/autosaves-{uid}.lock")
             if not lockfile.tryLock(0):# program created this still running
                 continue
             lockfile.unlock()
@@ -163,6 +171,10 @@ class AutosaveManager(QObject):
         return True
 
     def show_settings(self):
+        if not self.is_supported:
+            QMessageBox.information(self.window, "AutoSave not Supported",
+            "AutoSave features does not work on Flatpak. \nYou can use other formats like snap, deb package \nor pip-install to use this feature")
+            return
         dlg = AutosaveSettingsDialog(self.window)
         if dlg.exec()==QDialog.Accepted:
             enable, interval = dlg.getValues()
